Add files via upload
The goal of this competition is to identify breast cancer. I train my model with screening mammograms obtained from regular screening. This repository contains the code that I submitted during the competition in Kaggle. My work improving the automation of detection in screening mammography may enable radiologists to be more accurate and efficient, improving the quality and safety of patient care. It could also help reduce costs and unnecessary medical procedures.